cambios esteticos y funcionales se instalo bootstrap se agregaron las clases de bootstrap en los elementos html se implento la funcionalidad de agregar y quitar tutores a los alumnos

+++ database/migrations/2014_02_19_194135_create_datos_personales_table.php
@@ -14,7 +14,7 @@ class CreateDatosPersonalesTable extends Migration
     public function up()
     {
         Schema::create('datos_personales', function (Blueprint $table) {
-            $table->integer('id_dp')->unsigned();
+            $table->BigInteger('id_dp')->unsigned();
             $table->primary('id_dp');
 
             $table->char('tipo_documento', 4);
